Creates a class of tests to be run only if x11 libs are available
I started with the already existing test for the dualrenderer in the
internal viewer library. In the configure step the search for x11 libs
is added as we need to search for a running X-session in those tests. If
no x11 libs are available by accident, the tests are already skipped in
the Makefile of the tests-directory.

If the libs are available but no session is running, the test are
automatically skipped in the binary files of the tests (see
iv_dualrenderer_test.cc as a first example).