Removes useless document_icon_dir option
This option, by default a path to a directory with xmp files, could be
set by the user when the layout was set to 'MIME type icons'. Setting
this string value to a non-existing path didn't change anything in the
file panes view. This option seemes to be not fully implemented and
therefore I decided to remove this option completely, and all its
related lines of code.