Adds migration of custom colours into GSettings
As colours are represented by a string of special form, i.e.
'#rrrrggggbbbb', this commit for the migration of custom colours is a
bit longer than the others. I introduced some new functions for
validation of a colour string, so that miss-configuration of the
GSettings values for colour keys should not be possible so easily.