WIP: Remove the call of gnome_vfs_directory_list_load()
For doing that I had to change and migrate many methods as we are moving away from
using GnomeVFSFileInfo, instead we are using GFileInfo now. But the transition is
not finished yet and the commit here is only preliminary work.

For example, Gnome Commander can be started at the moment, but the file names in
the left and right panes are not shown. This and more has to be fixed later on.

When jumping to the upper directory, the programm crashes.