Respect the users choice what to do if a file which should be copied already exists
For this to work I first added a new entry in the Gnome Commander options dialog for
automatically renaming a copied file if it's filename already exists in the target
directory (see gnome-cmd-types.h).

The selection is used as a preselected radio button in the copy preparation dialog
in the GUI. Technically, the selection is forwarded through the overwriteMode member
of the XferData object over to the methods in gnome-cmd-xfer.cc.

In this file I also introduced a couple of new functions, which are actually only
extracts of source code from other places. But in these functions itself, the
overwriteMode is also processed, to decide what to do if a file would be overwritten
by the copy.