Remove the default Samba connection button from the device list
The reason I remove this is that I was not able in my local Samba tests
to get a list of browsable folders after clicking on the SMB connection
button. Instead, I could connect to a local Samba folder through the
remote connection setup dialog and I could also browse through that folder.

To reduce user confusion I decided to remove that button then...