Add signal handling for adding and removing a GMount
This is actually important for the integration of Gnome Commander into the desktop,
because connections to remote resources, i.e. GMounts, can also be established or closed
with other programs than Gnome Commander. These connections should then be integrated
in the programm, for example by closing the connection in the file pane when it the
connection itself was closed externally.