Don't use g_file_measure_disk_usage() before calling g_file_delete, finally fixing #59
This is needed when working with remote directorys. Instead, we are just trying
to delete the file/directory. If it is a non-empty directory, the error type
G_IO_ERROR_NOT_EMPTY is received, after which we can step into the non-empty
directory and recursively delete everything there.

Furthermore, I cleaned up the code a bit by splitting up the original function.